perf(route-registry): avoid heap alloc on uri split
The Trie used for the Route Registry is a multi-level map that
contains information for routes and route paths, and their endpoints.

Lookups and modifications on this Trie are always done based on the
provided URI, which is then split into its path segments. This split
always results in 2 parts. string.SplitN() will allocate the result
array for the results. As we know the number of expected results, and
returning slices of existing arrays does not incur a heap allocation,
we can instead return the slices "before" the separator and "after"
the separator, if there is one.

Go 1.21 now has a new `strings` package with the function `Cut` that
does exactly that.