fix(slog): only call complex log field population when level is enabled
Some of the log messages pass on complex data to be logged.
When the log level is not enabled, the input data is already
computed but then discarded by the logger, after checking
the active log level.

This check looks redundant but avoids many allocations and
processing that are not needed.

A nicer solution would be lazy evaluation, e.g. via callback
closures, but log/slog does not support those.