test: don't use fmt.Print* in tests
When outputting information in tests, Ginkgo recommend to use the
GinkgoWriter as that is able to capture the output and only print it in
case of failure.

This commit switches the obvious cases to use the GinkgoWriter instaed
of fmt.Print*. There are other indirect occurrences (for example by the
net/http package) but those can't easily be replaced.

See: https://onsi.github.io/ginkgo/#logging-output